better unit test
