+++ minimum_dependency_generator/tests/conftest.py
@@ -44,3 +44,51 @@ def setuptools():
 @pytest.fixture(scope="session", autouse=True)
 def p_ytest_dep():
     return "pytest >= 5.2.0"
+
+
+@pytest.fixture(scope="session", autouse=True)
+def cfg_str(dask_dep, pandas_dep, woodwork_dep, numpy_lower, ploty_dep, numpy_upper, p_ytest_dep):
+    setup_cfg_str = f'''\
+    [metadata]
+    name = example_package
+
+    [options]
+    install_requires =
+        {dask_dep}
+        {pandas_dep}
+        {woodwork_dep}
+        {numpy_lower}
+
+    [options.extras_require] =
+    dev =
+        {ploty_dep}
+        {numpy_upper}
+    test =
+        {p_ytest_dep}
+    '''
+    return setup_cfg_str
+
+@pytest.fixture(scope="session", autouse=True)
+def toml_cfg(pandas_dep, woodwork_dep, numpy_upper, ploty_dep, p_ytest_dep, dask_dep, numpy_lower):
+    pyproject_str = f'''\
+    [project]
+    name = "example_package"
+    requires-python = ">=3.7,<3.10"
+    dependencies = [
+        "{pandas_dep}",
+        "{woodwork_dep}",
+        "{numpy_upper}",
+    ]
+
+    [project.optional-dependencies]
+    test = [
+        "{ploty_dep}",
+        "{p_ytest_dep}",
+    ]
+    dev = [
+        "{dask_dep}",
+        "{numpy_lower}",
+    ]
+    '''
+    return pyproject_str
+
