+++ minimum_dependency_generator/tests/test_pyproject_toml.py
@@ -1,40 +1,47 @@
 import tempfile
 
+import pytest
+
 from ..minimum_dependency_generator import generate_min_requirements
-from .test_setup_cfg import verify_min_reqs_cfg_toml
 
-
-def test_with_pyproject_toml(
-    dask_dep, pandas_dep, woodwork_dep, numpy_lower, ploty_dep, numpy_upper, p_ytest_dep
+@pytest.mark.parametrize(
+    "file_prefix,file_extension,options",
+    [('setup', 'cfg', 'install_requires' ), ('pyproject', 'toml', 'dependencies')],
+)
+def test_with_toml_cfg(
+    file_prefix, file_extension, options, cfg_str, toml_cfg
 ):
-    pyproject_str = f'''\
-    [project]
-    name = "example_package"
-    requires-python = ">=3.7,<3.10"
-    dependencies = [
-        "{pandas_dep}",
-        "{woodwork_dep}",
-        "{numpy_upper}",
-    ]
-
-    [project.optional-dependencies]
-    test = [
-        "{ploty_dep}",
-        "{p_ytest_dep}",
-    ]
-    dev = [
-        "{dask_dep}",
-        "{numpy_lower}",
-    ]
-    '''
+    file_str = toml_cfg
+    if file_prefix == 'setup':
+        file_str = cfg_str
     with tempfile.NamedTemporaryFile(
-        mode="w", suffix=".toml", prefix="pyproject"
+        mode="w", suffix="." + file_extension, prefix=file_prefix
     ) as pyproject_file:
-        pyproject_file.write(pyproject_str)
+        pyproject_file.write(file_str)
         pyproject_file.flush()
 
         paths = [pyproject_file.name]
-        options = ['dependencies']
+        options = [options]
         extra_requires = ['test dev']
         min_requirements = generate_min_requirements(paths, options, extra_requires)
     verify_min_reqs_cfg_toml(min_requirements)
+
+def verify_min_reqs_cfg_toml(min_requirements):
+    assert '-r' not in min_requirements
+    assert '.txt' not in min_requirements
+    assert 'core-requirements.txt' not in min_requirements
+    min_requirements = min_requirements.split('\n')
+    assert min_requirements[-1] == ''
+    min_requirements = min_requirements[:-1]
+    expected_min_reqs = [
+        "dask[dataframe]==2.30.0",
+        "numpy==1.15.4",
+        "pandas==0.24.1",
+        "plotly==4.14.0",
+        "pytest==5.2.0",
+        "woodwork==0.0.11",
+    ]
+    expected_min_reqs = sorted(expected_min_reqs)
+    assert len(min_requirements) == len(expected_min_reqs)
+    for idx, min_req in enumerate(min_requirements):
+        assert expected_min_reqs[idx] == min_req.strip()
